Add files via upload
The study concerns the financial data of the company PODILATO.EN. The data is fictional and was created solely for training and practice purposes of certain tools, such as Excel, SQL, Python, PowerBI, and Tableau. It does not reflect real business data. The analysis was conducted in the MS Excel environment, and its graphical representation was completed using PowerBI. The entire analysis consists of a txt file (read.me), the Excel file (PODILATO.EN), the PowerBI file (PODILATO.EN), and the PDF (PODILATO.EN).

Additional information about the methodology and steps of this mock analysis can be found in the read.me file